test: split aws-sdk/v3/dynamodb into two logical units
